Account structure reorganization
