+++ accounts/templates/registration/login.html
@@ -15,7 +15,7 @@ Jesteś zalogowany jako <b>{{ user }}</b>.
 </form>
 <br>
 Nie masz konta? Załuż je!
-<form method="get" action="{% url 'website:register_new_user' %}">
+<form method="get" action="{% url 'accounts:register_new_user' %}">
 
     <input type="submit" value="Zarejestruj!">
 </form>
