+++ src/urls.py
@@ -22,5 +22,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('website.urls')),
     path('', include('django.contrib.auth.urls')),
+    path('accounts/', include('accounts.urls')),
 
 ]
